apis/tenancy: split up types.go
